+++ ImageOverlayRenewal/Data/TextureTransformMode.cs
@@ -1,0 +1,6 @@
+﻿namespace ImageOverlayRenewal.Data;
+
+public enum TextureTransformMode {
+    FlipVertical,
+    Transpose
+}