+++ ImageOverlayRenewal/UI/OptionsPanel.cs
@@ -2,11 +2,13 @@
 using System.IO;
 using ColossalFramework.IO;
 using ColossalFramework.UI;
+using CSLModsCommon;
 using CSLModsCommon.Localization;
 using CSLModsCommon.Logging;
 using CSLModsCommon.ToolButton;
 using CSLModsCommon.UI.Containers;
 using CSLModsCommon.UI.OptionsPanel;
+using ImageOverlayRenewal.Data;
 using ImageOverlayRenewal.Localization;
 using ImageOverlayRenewal.Managers;
 using ImageOverlayRenewal.Settings;
@@ -17,11 +19,11 @@ namespace ImageOverlayRenewal.UI;
 internal class OptionsPanel : OptionsPanelBase {
     private static readonly string PngDirectory = Application.platform == RuntimePlatform.OSXEditor || Application.platform == RuntimePlatform.OSXPlayer ? GetOSXDirectory() : DataLocation.gameContentPath;
 
-    private ModSetting _modModSetting;
+    private ModSetting _modSetting;
 
     protected override void CacheManagers() {
         base.CacheManagers();
-        _modModSetting = _settingManager.GetSetting<ModSetting>();
+        _modSetting = _settingManager.GetSetting<ModSetting>();
     }
 
     protected override void FillDebugPage(ScrollContainer page) {
@@ -32,7 +34,7 @@ internal class OptionsPanel : OptionsPanelBase {
     protected override InGameToolManagerBase GetInGameToolManager() => _domain.GetOrCreateManager<InGameToolButtonManager>();
 
     protected override void FillGeneralPage(ScrollContainer page) {
-        AddSection(page, Translations.LoadSettings).AddToggleSwitch(_modModSetting.ShowReloadResults, Translations.OptionPanel_ShowReloadResults, null, (_, b) => _modModSetting.ShowReloadResults = b);
+        AddSection(page, Translations.LoadSettings).AddToggleSwitch(_modSetting.ShowReloadResults, Translations.OptionPanel_ShowReloadResults, null, (_, b) => _modSetting.ShowReloadResults = b);
 
         var pngSection = AddSection(page, Translations.OptionPanel_PNGOptions);
         pngSection.AddStringField(Translations.OptionPanel_PNGFilePath, null, PngDirectory, fieldWidth:
@@ -43,15 +45,21 @@ internal class OptionsPanel : OptionsPanelBase {
             .Direction = FlexDirection.Column;
         pngSection.AddButton(Translations.OptionPanel_OpenPNGDirectory, null, Translations.OptionPanel_OpenPNGDirectory, null, 30, _ => Process.Start(PngDirectory));
 
+        var textureTransformModeRadioGroupCard = pngSection.AddEnumRadioGroup(Translations.TextureTransformMode, Translations.TextureTransformModeDescription, _modSetting.TransformMode, value => _modSetting.TransformMode = value, mode => mode switch {
+            TextureTransformMode.FlipVertical => Translations.FlipVertical,
+            _ => Translations.Transpose,
+        });
+        textureTransformModeRadioGroupCard.isEnabled = _domain.GetModManager().CurrentMode == GameMode.MainMenu;
+
         AddInGameToolButtonSection(value => _domain.GetOrCreateManager<InGameToolButtonManager>().OnButtonStatuesChanged(value));
     }
 
     protected override void FillKeyBindingPage(ScrollContainer page) {
         base.FillKeyBindingPage(page);
         var keyBindingSection = AddSection(page);
-        keyBindingSection.AddKeyBinding(_modModSetting.ControlPanelToggleKeyBinding, SharedTranslations.ToggleControlPanel, SharedTranslations.ToggleControlPanelDescription);
-        keyBindingSection.AddKeyBinding(_modModSetting.ShowImageKeyBinding, Translations.ControlPanel_ShowImage);
-        keyBindingSection.AddKeyBinding(_modModSetting.LoopImageKeyBinding, Translations.ControlPanel_LoopImage);
+        keyBindingSection.AddKeyBinding(_modSetting.ControlPanelToggleKeyBinding, SharedTranslations.ToggleControlPanel, SharedTranslations.ToggleControlPanelDescription);
+        keyBindingSection.AddKeyBinding(_modSetting.ShowImageKeyBinding, Translations.ControlPanel_ShowImage);
+        keyBindingSection.AddKeyBinding(_modSetting.LoopImageKeyBinding, Translations.ControlPanel_LoopImage);
     }
 
     private static string GetOSXDirectory() {